separate unit-tests into a dedicated file per each test category
