+++ ppl-spark-integration/src/main/java/org/opensearch/sql/ppl/utils/ComparatorTransformer.java
@@ -3,9 +3,17 @@ package org.opensearch.sql.ppl.utils;
 import org.apache.spark.sql.catalyst.expressions.BinaryComparison;
 import org.apache.spark.sql.catalyst.expressions.EqualTo;
 import org.apache.spark.sql.catalyst.expressions.Expression;
+import org.apache.spark.sql.catalyst.expressions.GreaterThan;
+import org.apache.spark.sql.catalyst.expressions.GreaterThanOrEqual;
+import org.apache.spark.sql.catalyst.expressions.LessThan;
+import org.apache.spark.sql.catalyst.expressions.LessThanOrEqual;
+import org.apache.spark.sql.catalyst.expressions.Not;
+import org.apache.spark.sql.catalyst.expressions.Predicate;
 import org.opensearch.sql.ast.expression.Compare;
 import org.opensearch.sql.expression.function.BuiltinFunctionName;
 import org.opensearch.sql.ppl.CatalystPlanContext;
+
+import static com.amazonaws.services.mturk.model.Comparator.NotEqualTo;
 
 /**
  * Transform the PPL Logical comparator into catalyst comparator
@@ -15,7 +23,7 @@ public interface ComparatorTransformer {
      * comparator expression builder building a catalyst binary comparator from PPL's compare logical step
      * @return
      */
-    static BinaryComparison comparator(Compare expression, CatalystPlanContext context) {
+    static Predicate comparator(Compare expression, CatalystPlanContext context) {
         if (BuiltinFunctionName.of(expression.getOperator()).isEmpty())
             throw new IllegalStateException("Unexpected value: " + BuiltinFunctionName.of(expression.getOperator()));
 
@@ -274,15 +282,15 @@ public interface ComparatorTransformer {
             case EQUAL:
                 return new EqualTo(left,right);
             case NOTEQUAL:
-                break;
+                return new Not(new EqualTo(left,right));
             case LESS:
-                break;
+                return new LessThan(left,right);
             case LTE:
-                break;
+                return new LessThanOrEqual(left,right);
             case GREATER:
-                break;
+                return new GreaterThan(left,right);
             case GTE:
-                break;
+                return new GreaterThanOrEqual(left,right);
             case LIKE:
                 break;
             case NOT_LIKE:
