+++ ppl-spark-integration/src/test/scala/org/opensearch/flint/spark/ppl/PPLLogicalPlanSimpleTranslatorTestSuite.scala
@@ -11,7 +11,7 @@ import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.catalyst.analysis.{Analyzer, FunctionRegistry, TableFunctionRegistry, UnresolvedAttribute, UnresolvedRelation, UnresolvedStar, UnresolvedTable}
 import org.apache.spark.sql.catalyst.catalog.{CatalogStorageFormat, CatalogTable, CatalogTableType, ExternalCatalog, FunctionExpressionBuilder, FunctionResourceLoader, GlobalTempViewManager, SessionCatalog}
 import org.apache.spark.sql.catalyst.expressions.aggregate.{AggregateExpression, Average, Complete, Count, Max}
-import org.apache.spark.sql.catalyst.expressions.{Alias, And, Descending, Divide, EqualTo, Floor, GreaterThan, GreaterThanOrEqual, LessThan, Like, Literal, NamedExpression, SortOrder, UnixTimestamp}
+import org.apache.spark.sql.catalyst.expressions.{Alias, And, Descending, Divide, EqualTo, Floor, GreaterThan, GreaterThanOrEqual, LessThan, LessThanOrEqual, Like, Literal, NamedExpression, Not, SortOrder, UnixTimestamp}
 import org.apache.spark.sql.catalyst.parser.{CatalystSqlParser, ParserInterface}
 import org.apache.spark.sql.catalyst.plans.logical.{Aggregate, Filter, Limit, LocalRelation, LogicalPlan, Project, Sort, Union}
 import org.apache.spark.sql.internal.SQLConf
@@ -23,217 +23,13 @@ import org.opensearch.sql.ppl.{CatalystPlanContext, CatalystQueryPlanVisitor}
 import org.scalatest.matchers.should.Matchers
 import org.scalatestplus.mockito.MockitoSugar.mock
 
-class PPLLogicalPlanTranslatorTestSuite
+class PPLLogicalPlanSimpleTranslatorTestSuite
   extends SparkFunSuite
     with Matchers {
 
   private val planTrnasformer = new CatalystQueryPlanVisitor()
   private val pplParser = new PPLSyntaxParser()
-
-  test("test simple search with only one table and no explicit fields (defaults to all fields)") {
-    // if successful build ppl logical plan and translate to catalyst logical plan
-    val context = new CatalystPlanContext
-    val logPlan = planTrnasformer.visit(plan(pplParser, "source=table", false), context)
-
-    val projectList: Seq[NamedExpression] = Seq(UnresolvedStar(None))
-    val expectedPlan = Project(projectList, UnresolvedRelation(Seq("table")))
-    assertEquals(expectedPlan, context.getPlan)
-    assertEquals(logPlan, "source=[table] | fields + *")
-
-  }
-
-  test("test simple search with schema.table and no explicit fields (defaults to all fields)") {
-    // if successful build ppl logical plan and translate to catalyst logical plan
-    val context = new CatalystPlanContext
-    val logPlan = planTrnasformer.visit(plan(pplParser, "source=schema.table", false), context)
-
-    val projectList: Seq[NamedExpression] = Seq(UnresolvedStar(None))
-    val expectedPlan = Project(projectList, UnresolvedRelation(Seq("schema", "table")))
-    assertEquals(expectedPlan, context.getPlan)
-    assertEquals(logPlan, "source=[schema.table] | fields + *")
-
-  }
-
-  test("test simple search with schema.table and one field projected") {
-    val context = new CatalystPlanContext
-    val logPlan = planTrnasformer.visit(plan(pplParser, "source=schema.table | fields A", false), context)
-
-    val projectList: Seq[NamedExpression] = Seq(UnresolvedAttribute("A"))
-    val expectedPlan = Project(projectList, UnresolvedRelation(Seq("schema", "table")))
-    assertEquals(expectedPlan, context.getPlan)
-    assertEquals(logPlan, "source=[schema.table] | fields + A")
-  }
-
-  test("test simple search with only one table with one field projected") {
-    val context = new CatalystPlanContext
-    val logPlan = planTrnasformer.visit(plan(pplParser, "source=table | fields A", false), context)
-
-    val projectList: Seq[NamedExpression] = Seq(UnresolvedAttribute("A"))
-    val expectedPlan = Project(projectList, UnresolvedRelation(Seq("table")))
-    assertEquals(expectedPlan, context.getPlan)
-    assertEquals(logPlan, "source=[table] | fields + A")
-  }
-
-  test("test simple search with only one table with one field literal filtered ") {
-    val context = new CatalystPlanContext
-    val logPlan = planTrnasformer.visit(plan(pplParser, "source=t a = 1 ", false), context)
-
-    val table = UnresolvedRelation(Seq("t"))
-    val filterExpr = EqualTo(UnresolvedAttribute("a"), Literal(1))
-    val filterPlan = Filter(filterExpr, table)
-    val projectList = Seq(UnresolvedStar(None))
-    val expectedPlan = Project(projectList, filterPlan)
-    assertEquals(expectedPlan, context.getPlan)
-    assertEquals(logPlan, "source=[t] | where a = 1 | fields + *")
-  }
-
-  test("test simple search with only one table with one field literal int equality filtered and one field projected") {
-    val context = new CatalystPlanContext
-    val logPlan = planTrnasformer.visit(plan(pplParser, "source=t a = 1  | fields a", false), context)
-
-    val table = UnresolvedRelation(Seq("t"))
-    val filterExpr = EqualTo(UnresolvedAttribute("a"), Literal(1))
-    val filterPlan = Filter(filterExpr, table)
-    val projectList = Seq(UnresolvedAttribute("a"))
-    val expectedPlan = Project(projectList, filterPlan)
-    assertEquals(expectedPlan, context.getPlan)
-    assertEquals(logPlan, "source=[t] | where a = 1 | fields + a")
-  }
-
-  test("test simple search with only one table with one field literal string equality filtered and one field projected") {
-    val context = new CatalystPlanContext
-    val logPlan = planTrnasformer.visit(plan(pplParser,  """source=t a = 'hi'  | fields a""", false), context)
-
-    val table = UnresolvedRelation(Seq("t"))
-    val filterExpr = EqualTo(UnresolvedAttribute("a"), Literal("'hi'"))
-    val filterPlan = Filter(filterExpr, table)
-    val projectList = Seq(UnresolvedAttribute("a"))
-    val expectedPlan = Project(projectList, filterPlan)
- 
-    assertEquals(expectedPlan,context.getPlan)
-    assertEquals(logPlan, "source=[t] | where a = 'hi' | fields + a")
-  }
-
-  test("test simple search with only one table with one field greater than  filtered and one field projected") {
-    val context = new CatalystPlanContext
-    val logPlan = planTrnasformer.visit(plan(pplParser, "source=t a > 1  | fields a", false), context)
-
-    val table = UnresolvedRelation(Seq("t"))
-    val filterExpr = EqualTo(UnresolvedAttribute("a"), Literal(1))
-    val filterPlan = Filter(filterExpr, table)
-    val projectList = Seq(UnresolvedAttribute("a"))
-    val expectedPlan = Project(projectList, filterPlan)
-    assertEquals(expectedPlan, context.getPlan)
-    assertEquals(logPlan, "source=[t] | where a > 1 | fields + a")
-  }
-
-  test("test simple search with only one table with one field greater than equal  filtered and one field projected") {
-    val context = new CatalystPlanContext
-    val logPlan = planTrnasformer.visit(plan(pplParser, "source=t a >= 1  | fields a", false), context)
-
-    val table = UnresolvedRelation(Seq("t"))
-    val filterExpr = EqualTo(UnresolvedAttribute("a"), Literal(1))
-    val filterPlan = Filter(filterExpr, table)
-    val projectList = Seq(UnresolvedAttribute("a"))
-    val expectedPlan = Project(projectList, filterPlan)
-    assertEquals(expectedPlan, context.getPlan)
-    assertEquals(logPlan, "source=[t] | where a >= 1 | fields + a")
-  }
-
-  test("test simple search with only one table with one field lower than filtered and one field projected") {
-    val context = new CatalystPlanContext
-    val logPlan = planTrnasformer.visit(plan(pplParser, "source=t a < 1  | fields a", false), context)
-
-    val table = UnresolvedRelation(Seq("t"))
-    val filterExpr = EqualTo(UnresolvedAttribute("a"), Literal(1))
-    val filterPlan = Filter(filterExpr, table)
-    val projectList = Seq(UnresolvedAttribute("a"))
-    val expectedPlan = Project(projectList, filterPlan)
-    assertEquals(expectedPlan, context.getPlan)
-    assertEquals(logPlan, "source=[t] | where a < 1 | fields + a")
-  }
-
-  test("test simple search with only one table with one field lower than equal filtered and one field projected") {
-    val context = new CatalystPlanContext
-    val logPlan = planTrnasformer.visit(plan(pplParser, "source=t a <= 1  | fields a", false), context)
-
-    val table = UnresolvedRelation(Seq("t"))
-    val filterExpr = EqualTo(UnresolvedAttribute("a"), Literal(1))
-    val filterPlan = Filter(filterExpr, table)
-    val projectList = Seq(UnresolvedAttribute("a"))
-    val expectedPlan = Project(projectList, filterPlan)
-    assertEquals(expectedPlan, context.getPlan)
-    assertEquals(logPlan, "source=[t] | where a <= 1 | fields + a")
-  }
-
-  test("test simple search with only one table with one field not equal filtered and one field projected") {
-    val context = new CatalystPlanContext
-    val logPlan = planTrnasformer.visit(plan(pplParser, "source=t a != 1  | fields a", false), context)
-
-    val table = UnresolvedRelation(Seq("t"))
-    val filterExpr = EqualTo(UnresolvedAttribute("a"), Literal(1))
-    val filterPlan = Filter(filterExpr, table)
-    val projectList = Seq(UnresolvedAttribute("a"))
-    val expectedPlan = Project(projectList, filterPlan)
-    assertEquals(expectedPlan, context.getPlan)
-    assertEquals(logPlan, "source=[t] | where a != 1 | fields + a")
-  }
-
-
-  test("test simple search with only one table with two fields projected") {
-    val context = new CatalystPlanContext
-    val logPlan = planTrnasformer.visit(plan(pplParser, "source=t | fields A, B", false), context)
-
-
-    val table = UnresolvedRelation(Seq("t"))
-    val projectList = Seq(UnresolvedAttribute("A"), UnresolvedAttribute("B"))
-    val expectedPlan = Project(projectList, table)
-    assertEquals(expectedPlan, context.getPlan)
-    assertEquals(logPlan, "source=[t] | fields + A,B")
-  }
-
-
-  test("Search multiple tables - translated into union call - fields expected to exist in both tables ") {
-    val context = new CatalystPlanContext
-    val logPlan = planTrnasformer.visit(plan(pplParser, "search source = table1, table2 | fields A, B", false), context)
-
-
-    val table1 = UnresolvedRelation(Seq("table1"))
-    val table2 = UnresolvedRelation(Seq("table2"))
-
-    val allFields1 = Seq(UnresolvedAttribute("A"), UnresolvedAttribute("B"))
-    val allFields2 = Seq(UnresolvedAttribute("A"), UnresolvedAttribute("B"))
-
-    val projectedTable1 = Project(allFields1, table1)
-    val projectedTable2 = Project(allFields2, table2)
-
-    val expectedPlan = Union(Seq(projectedTable1, projectedTable2), byName = true, allowMissingCol = true)
-
-    assertEquals(logPlan, "source=[table1, table2] | fields + A,B")
-    assertEquals(expectedPlan, context.getPlan)
-  }
-
-
-  test("Search multiple tables - translated into union call with fields") {
-    val context = new CatalystPlanContext
-    val logPlan = planTrnasformer.visit(plan(pplParser, "search source = table1, table2 | ", false), context)
-
-
-    val table1 = UnresolvedRelation(Seq("table1"))
-    val table2 = UnresolvedRelation(Seq("table2"))
-
-    val allFields1 = UnresolvedStar(None)
-    val allFields2 = UnresolvedStar(None)
-
-    val projectedTable1 = Project(Seq(allFields1), table1)
-    val projectedTable2 = Project(Seq(allFields2), table2)
-
-    val expectedPlan = Union(Seq(projectedTable1, projectedTable2), byName = true, allowMissingCol = true)
-
-    assertEquals(logPlan, "source=[table1, table2] | fields + *")
-    assertEquals(expectedPlan, context.getPlan)
-  }
-
+  
   test("Find What are the average prices for different types of properties") {
     val context = new CatalystPlanContext
     val logPlan = planTrnasformer.visit(plan(pplParser, "source = housing_properties | stats avg(price) by property_type", false), context)
@@ -620,61 +416,5 @@ class PPLLogicalPlanTranslatorTestSuite
     assertEquals(expectedPlan, context.getPlan)
     assertEquals(logPlan, "???")
 
-  }
-  
-  //  TODO - fix
-  test("Test Analyzer with Logical Plan") {
-    // Mock table schema and existence
-    val tableSchema = StructType(
-      List(
-        StructField("nonexistent_column", IntegerType),
-        StructField("another_nonexistent_column", IntegerType)
-      )
-    )
-    val catalogTable = CatalogTable(
-      identifier = TableIdentifier("nonexistent_table"),
-      tableType = CatalogTableType.MANAGED,
-      storage = CatalogStorageFormat.empty,
-      schema = tableSchema
-    )
-    val externalCatalog = mock[ExternalCatalog]
-    when(externalCatalog.tableExists("default", "nonexistent_table")).thenReturn(true)
-    when(externalCatalog.getTable("default", "nonexistent_table")).thenReturn(catalogTable)
-
-    // Mocking required components
-    val functionRegistry = mock[FunctionRegistry]
-    val tableFunctionRegistry = mock[TableFunctionRegistry]
-    val globalTempViewManager = mock[GlobalTempViewManager]
-    val functionResourceLoader = mock[FunctionResourceLoader]
-    val functionExpressionBuilder = mock[FunctionExpressionBuilder]
-    val hadoopConf = new Configuration()
-    val sqlParser = mock[ParserInterface]
-
-    val emptyCatalog = new SessionCatalog(
-      externalCatalogBuilder = () => externalCatalog,
-      globalTempViewManagerBuilder = () => globalTempViewManager,
-      functionRegistry = functionRegistry,
-      tableFunctionRegistry = tableFunctionRegistry,
-      hadoopConf = hadoopConf,
-      parser = sqlParser,
-      functionResourceLoader = functionResourceLoader,
-      functionExpressionBuilder = functionExpressionBuilder,
-      cacheSize = 1000,
-      cacheTTL = 0L
-    )
-
-
-    val analyzer = new Analyzer(emptyCatalog)
-
-    // Create a sample LogicalPlan
-    val invalidLogicalPlan = Project(
-      Seq(Alias(UnresolvedAttribute("undefined_column"), "alias")()),
-      LocalRelation()
-    )
-    // Analyze the LogicalPlan
-    val resolvedLogicalPlan: LogicalPlan = analyzer.execute(invalidLogicalPlan)
-
-    // Assertions to check the validity of the analyzed plan
-    assert(resolvedLogicalPlan.resolved)
   }
 }
